+++ docs/changelog.md
@@ -12,5 +12,6 @@ Version 1.1.0 ()
 
 * Added Template mapping closure in `app/_configs/default.php` (`router.template`).
 * Added docs for Configuration and Models.
+* Added `ini_set('display_errors', 'on');` so the debug parameters in the config are always working
+* Added `Factory::onload` so you don't have to set a template path if you create an instance of the \Morrow\Views\Serpent class
 * Removed `composer.lock`.
-* Added `ini_set('display_errors', 'on');`